codecrafters submit [skip ci]

+++ src/Common/RespBuilder.cs
@@ -16,22 +16,22 @@ public static class RespBuilder
 
         return sb.ToString();
     }
-    
+
     public static string BuildRespBulkString(string value)
     {
         return $"${value.Length}\r\n{value}\r\n";
     }
-    
-    public static string BuildRespInteger(int value)
+
+    public static string BuildRespInteger(long value)
     {
         return $":{value}\r\n";
     }
-    
+
     public static string BuildRespError(string value)
     {
         return $"-ERR {value}\r\n";
     }
-    
+
     public static byte[] AsBytes(this string value)
     {
         return Encoding.UTF8.GetBytes(value);
